Tentatively finish Context conversion.
Needs proper testing, which I'll do once Go 1.7 is released.

Decided to just drop backward compatibility as the code was
becoming complex. I will keep the originals style on master and
publish this on a new branch when it's ready, and use gopkg.in to
provide easy access to it.

See #30